Phase 0: Extract BolusManager from CommService
Move bolus confirmation logic (~180 lines) into pump/BolusManager.kt:
- BolusRequestSource enum
- confirmBolusRequest() with notification and auto-approve logic
- scheduleAutoApprove() / cancelAutoApproveStatic()
- resetBolusPrefs()
- confirmBolusRequestBaseNotification() helper

CommService now delegates bolus operations to BolusManager, reducing
CommService from ~912 to ~730 lines. This completes Phase 0 of the
watch-as-host refactor plan.
