+++ mobile/src/main/java/com/jwoglom/controlx2/CommService.kt
@@ -4,15 +4,12 @@ import android.app.Notification
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.app.PendingIntent
-import android.app.PendingIntent.FLAG_IMMUTABLE
-import android.app.PendingIntent.FLAG_ONE_SHOT
 import android.app.Service
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
 import android.content.SharedPreferences
 import android.content.pm.ServiceInfo.FOREGROUND_SERVICE_TYPE_CONNECTED_DEVICE
-import android.media.RingtoneManager
 import android.os.Handler
 import android.os.HandlerThread
 import android.os.Looper
@@ -26,26 +23,25 @@ import com.jwoglom.controlx2.db.historylog.HistoryLogRepo
 import com.jwoglom.controlx2.messaging.MessageBusFactory
 import com.jwoglom.controlx2.presentation.util.ShouldLogToFile
 import com.jwoglom.controlx2.pump.BleChangeReceiver
+import com.jwoglom.controlx2.pump.BolusManager
 import com.jwoglom.controlx2.pump.CommServiceCallbacks
 import com.jwoglom.controlx2.pump.PumpCommHandler
 import com.jwoglom.controlx2.pump.PumpCommState
 import com.jwoglom.controlx2.pump.PumpFinderCommHandler
 import com.jwoglom.controlx2.pump.PumpSession
 import com.jwoglom.controlx2.shared.CommServiceCodes
-import com.jwoglom.controlx2.shared.MessagePaths
 import com.jwoglom.controlx2.shared.InitiateConfirmedBolusSerializer
+import com.jwoglom.controlx2.shared.MessagePaths
 import com.jwoglom.controlx2.shared.PumpMessageSerializer
 import com.jwoglom.controlx2.shared.messaging.MessageBus
 import com.jwoglom.controlx2.shared.messaging.MessageBusSender
 import com.jwoglom.controlx2.shared.messaging.MessageListener
 import com.jwoglom.controlx2.shared.util.setupTimber
 import com.jwoglom.controlx2.shared.util.shortTime
-import com.jwoglom.controlx2.shared.util.twoDecimalPlaces
 import com.jwoglom.controlx2.util.AppVersionCheck
 import com.jwoglom.pumpx2.pump.PumpState
 import com.jwoglom.pumpx2.pump.messages.bluetooth.Characteristic
 import com.jwoglom.pumpx2.pump.messages.builders.CurrentBatteryRequestBuilder
-import com.jwoglom.pumpx2.pump.messages.helpers.Bytes
 import com.jwoglom.pumpx2.pump.messages.models.ApiVersion
 import com.jwoglom.pumpx2.pump.messages.models.InsulinUnit
 import com.jwoglom.pumpx2.pump.messages.models.KnownApiVersion
@@ -57,7 +53,6 @@ import com.jwoglom.pumpx2.pump.messages.request.currentStatus.InsulinStatusReque
 import com.jwoglom.pumpx2.pump.messages.response.currentStatus.ControlIQIOBResponse
 import com.jwoglom.pumpx2.pump.messages.response.currentStatus.CurrentBatteryAbstractResponse
 import com.jwoglom.pumpx2.pump.messages.response.currentStatus.InsulinStatusResponse
-import com.jwoglom.pumpx2.shared.Hex
 import com.welie.blessed.BluetoothPeripheral
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -88,6 +83,8 @@ class CommService : Service(), CommServiceCallbacks {
     internal fun simulateConnectedPump(peripheral: BluetoothPeripheral) {
         pumpCommHandler!!.simulateConnectedPump(peripheral)
     }
+
+    private lateinit var bolusManager: BolusManager
 
     private var serviceStatusAcknowledged = false
     private val serviceStatusTask = object : Runnable {
@@ -207,6 +204,8 @@ class CommService : Service(), CommServiceCallbacks {
             }
         })
 
+        bolusManager = BolusManager(applicationContext, ::sendWearCommMessage)
+
         // Initialize and start HTTP Debug API service
         httpDebugApiService = HttpDebugApiService(applicationContext)
         httpDebugApiService?.getCurrentPumpDataCallback = { getCurrentPumpDataJson() }
@@ -311,16 +310,14 @@ class CommService : Service(), CommServiceCallbacks {
                 sendPumpPairingMessage()
             }
             MessagePaths.TO_SERVER_BOLUS_REQUEST_WEAR -> {
-                // removed: initialized check
-                confirmBolusRequest(PumpMessageSerializer.fromBytes(data) as InitiateBolusRequest, BolusRequestSource.WEAR)
+                bolusManager.confirmBolusRequest(PumpMessageSerializer.fromBytes(data) as InitiateBolusRequest, BolusManager.BolusRequestSource.WEAR)
             }
             MessagePaths.TO_SERVER_BOLUS_REQUEST_PHONE -> {
-                // removed: initialized check
-                confirmBolusRequest(PumpMessageSerializer.fromBytes(data) as InitiateBolusRequest, BolusRequestSource.PHONE)
+                bolusManager.confirmBolusRequest(PumpMessageSerializer.fromBytes(data) as InitiateBolusRequest, BolusManager.BolusRequestSource.PHONE)
             }
             MessagePaths.TO_SERVER_BOLUS_CANCEL -> {
                 Timber.i("bolus state cancelled")
-                resetBolusPrefs(this)
+                bolusManager.resetBolusPrefs()
             }
             MessagePaths.TO_SERVER_INITIATE_CONFIRMED_BOLUS -> {
                 // removed: initialized check
@@ -640,151 +637,10 @@ class CommService : Service(), CommServiceCallbacks {
     }
 
 
-    private var bolusNotificationId: Int = 1000
-
-    enum class BolusRequestSource(val id: String) {
-        PHONE("phone"),
-        WEAR("wear")
-    }
-    private fun confirmBolusRequest(request: InitiateBolusRequest, source: BolusRequestSource) {
-        val units = twoDecimalPlaces(InsulinUnit.from1000To1(request.totalVolume))
-        Timber.i("confirmBolusRequest $units: $request")
-        bolusNotificationId++
-        prefs(this)?.edit()
-            ?.putString("initiateBolusRequest", Hex.encodeHexString(PumpMessageSerializer.toBytes(request)))
-            ?.putString("initiateBolusSecret", Hex.encodeHexString(Bytes.getSecureRandom10Bytes()))
-            ?.putString("initiateBolusSource", source.id)
-            ?.putLong("initiateBolusTime", Instant.now().toEpochMilli())
-            ?.putInt("initiateBolusNotificationId", bolusNotificationId)
-            ?.commit().let {
-                if (it != true) {
-                    Timber.e("synchronous preference write failed in confirmBolusRequest")
-                }
-            }
-
-        fun getRejectIntent(): PendingIntent {
-            val rejectIntent =
-                Intent(applicationContext, BolusNotificationBroadcastReceiver::class.java).apply {
-                    putExtra("action", "REJECT")
-                }
-            return PendingIntent.getBroadcast(
-                this,
-                2000,
-                rejectIntent,
-                FLAG_IMMUTABLE or FLAG_ONE_SHOT
-            )
-        }
-
-        fun getConfirmIntent(): PendingIntent {
-            val confirmIntent =
-                Intent(applicationContext, BolusNotificationBroadcastReceiver::class.java).apply {
-                    putExtra("action", "INITIATE")
-                    putExtra("request", PumpMessageSerializer.toBytes(request))
-                }
-            return PendingIntent.getBroadcast(
-                this,
-                2001,
-                confirmIntent,
-                FLAG_IMMUTABLE or FLAG_ONE_SHOT
-            )
-        }
-
-        val minNotifyThreshold = Prefs(this).bolusConfirmationInsulinThreshold()
-        val autoApproveTimeout = if (source == BolusRequestSource.WEAR)
-            Prefs(this).wearBolusAutoApproveTimeoutSeconds() else 0
-        sendWearCommMessage(MessagePaths.TO_CLIENT_BOLUS_MIN_NOTIFY_THRESHOLD, "$minNotifyThreshold".toByteArray())
-        sendWearCommMessage(MessagePaths.TO_CLIENT_WEAR_AUTO_APPROVE_TIMEOUT, "$autoApproveTimeout".toByteArray())
-
-        if (InsulinUnit.from1000To1(request.totalVolume) >= minNotifyThreshold || minNotifyThreshold == 0.0) {
-            Timber.i("Requesting permission for bolus because $units >= minNotifyThreshold=$minNotifyThreshold")
-
-            val builder = confirmBolusRequestBaseNotification(
-                this,
-                "Bolus Request",
-                if (autoApproveTimeout > 0) "$units units. Auto-approving in ${autoApproveTimeout}s unless canceled."
-                else "$units units. Press Confirm to deliver."
-            )
-                .setSound(RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION))
-                .setVibrate(longArrayOf(500L, 500L, 500L, 500L, 500L, 500L, 500L, 500L, 500L, 500L))
-
-            builder.addAction(R.drawable.decline, "Reject", getRejectIntent())
-
-            builder.addAction(R.drawable.bolus_icon, "Confirm ${units}u", getConfirmIntent())
-
-            val notif = builder.build()
-            Timber.i("bolus notification $bolusNotificationId $builder $notif")
-            makeNotif(bolusNotificationId, notif)
-
-            // Broadcast to phone UI for in-app dialog
-            sendWearCommMessage(MessagePaths.TO_SERVER_BOLUS_CONFIRM_DIALOG,
-                "${Hex.encodeHexString(PumpMessageSerializer.toBytes(request))}|${source.id}|$autoApproveTimeout".toByteArray())
-
-            // Schedule auto-approve if timeout is set
-            if (autoApproveTimeout > 0) {
-                scheduleAutoApprove(autoApproveTimeout, getConfirmIntent())
-            }
-        } else {
-            Timber.i("Sending immediate bolus request because $units is less than minNotifyThreshold=$minNotifyThreshold")
-
-            val confirmIntent =
-                Intent(applicationContext, BolusNotificationBroadcastReceiver::class.java).apply {
-                    putExtra("action", "INITIATE")
-                    putExtra("request", PumpMessageSerializer.toBytes(request))
-                }
-            val confirmPendingIntent = PendingIntent.getBroadcast(
-                this,
-                2001,
-                confirmIntent,
-                FLAG_IMMUTABLE or FLAG_ONE_SHOT
-            )
-            // wait to avoid prefs not being saved
-            Thread.sleep(250)
-            confirmPendingIntent.send()
-        }
-    }
-
-    private fun scheduleAutoApprove(timeoutSeconds: Int, confirmIntent: PendingIntent) {
-        cancelAutoApproveStatic(this)
-        val handler = Handler(Looper.getMainLooper())
-        _autoApproveHandler = handler
-        val runnable = Runnable {
-            Timber.i("Auto-approving bolus after ${timeoutSeconds}s timeout")
-            try {
-                confirmIntent.send()
-            } catch (e: PendingIntent.CanceledException) {
-                Timber.w("Auto-approve PendingIntent already cancelled (bolus was manually handled)")
-            }
-            _autoApproveRunnable = null
-            _autoApproveHandler = null
-        }
-        _autoApproveRunnable = runnable
-        handler.postDelayed(runnable, timeoutSeconds * 1000L)
-    }
-
     companion object {
-        @Volatile
-        private var _autoApproveHandler: Handler? = null
-        @Volatile
-        private var _autoApproveRunnable: Runnable? = null
-
         fun cancelAutoApproveStatic(context: Context?) {
-            _autoApproveRunnable?.let { runnable ->
-                _autoApproveHandler?.removeCallbacks(runnable)
-                Timber.i("Auto-approve timer cancelled")
-            }
-            _autoApproveRunnable = null
-            _autoApproveHandler = null
+            BolusManager.cancelAutoApproveStatic(context)
         }
-    }
-
-    private fun makeNotif(id: Int, notif: Notification) {
-        val notificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager;
-        notificationManager.notify(id, notif)
-    }
-
-    private fun cancelNotif(id: Int) {
-        val notificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager;
-        notificationManager.cancel(id)
     }
 
     override fun onDestroy() {
@@ -863,13 +719,6 @@ class CommService : Service(), CommServiceCallbacks {
         return context.getSharedPreferences("WearX2", MODE_PRIVATE)
     }
 
-    private fun resetBolusPrefs(context: Context) {
-        prefs(context)?.edit()
-            ?.remove("initiateBolusRequest")
-            ?.remove("initiateBolusTime")
-            ?.apply()
-    }
-
     private fun triggerAppReload(context: Context) {
         val packageManager = context.packageManager
         val intent = packageManager.getLaunchIntentForPackage(context.packageName)
@@ -878,36 +727,4 @@ class CommService : Service(), CommServiceCallbacks {
         context.startActivity(mainIntent)
         Runtime.getRuntime().exit(0)
     }
-}
-
-
-
-fun confirmBolusRequestBaseNotification(context: Context?, title: String, text: String): NotificationCompat.Builder {
-    val notificationChannelId = "Confirm Bolus"
-
-    val notificationManager = context?.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager;
-    val channel = NotificationChannel(
-        notificationChannelId,
-        "Confirm Bolus",
-        NotificationManager.IMPORTANCE_HIGH
-    ).let {
-        it.description = "Confirm Bolus"
-        it
-    }
-    notificationManager.createNotificationChannel(channel)
-
-    //val intent = Intent(this, MainActivity::class.java)
-    //val pendingIntent: PendingIntent = PendingIntent.getActivity(this, 0, intent, FLAG_IMMUTABLE)
-
-    return NotificationCompat.Builder(
-        context,
-        notificationChannelId
-    )
-        .setContentTitle(title)
-        .setContentText(text)
-        .setSmallIcon(R.drawable.bolus_icon)
-        .setTicker(title)
-        .setPriority(NotificationCompat.PRIORITY_MAX) // for under android 26 compatibility
-        .setAutoCancel(true)
-
 }
